requirements.txt
click
fastapi
uvicorn
docker
PyYAML
psutil
lxc
pihole6api
yoyo-migrations
psycopg2-binary
